Changes to support using the chrome.alarms API for > 30000 flipWait values
The changes in this commit represent a significant change to the way the extension works under-the-hood.
Previously the service worker was kept alive via an unsupported method, which allowed the old code to continue to work as a Manifest Version 3 extension.
However, this technique can be considered a hack. This commit attempts to embrace MV3 Service Workers by using them as they are supposed to be used. In this case, waking up the Service Worker using the chrome.alarms API.

There is logic added to determine whether the chrome.alarms API will be used (for flipWait values > 30000) or whether setInterval will be used (for flipWait values <= 30000). The reasons for this are two-fold:
* the chrome.alarms API minimum alarm interval is 30 seconds (thus we use setInterval)
* Service Workers become idle after 30 seconds (thus we use chome.alarms)  See here for more info on this: https://chromium.googlesource.com/chromium/src/+/main/docs/security/service-worker-security-faq.md#do-service-workers-live-forever

To support this, there are numerous changes to the Javascript Carousel class.

There are also some other noteworthy changes:
* options.html - Flip wait field - change the unit of measurement from milliseconds to seconds
               - add Reload interval field to set how frequently a browser refresh is done. If this is set to 0 then a refresh won't be done. This option's unit of measurement is minutes
               - change Save button message to prompt the user to restart the browser for settings changes to take effect
* background.js has been renamed to service_worker.js to accurately reflect what it is, as per the Google MV3 migration docs
* private class methods have been prefixed with an underscore for clarity