service_worker.js: clear redundant alarms unconditionally. Move automatic start evaluation out of orchestrate method.
Previously there were three calls to chrome.alarms.getAll & the _clear_redundant_alarms method, which was inefficient.
Also, these calls only ran if automatic start was enabled, which meant that alarms (& hence tab flipping/reloading) might undesirably fire even with automatic start disabled.
This commit replaces the three conditional calls with one unconditional call, and the _clear_redundant_alarms method now accepts an array of required alarms as the second argument, so that all other, redundant alarms can be cleared in one go.

There are also changes to where the automatic start value is evaluated. This was previously in the orchestrate method but has been moved closer to where the carouselling code lives, so that required alarms are still created, but won't initiate carouselling or reload if automatic start is disabled.